Made puppeteering work correctly. Tests run fast now, assuming you've run them once!

+++ features/support/grails/application.rb
@@ -11,12 +11,13 @@ module Grails
       # We need something File-like
       @root = File.open(directory) unless directory.respond_to?(:path)
       
-      # One day it'd be nice to load existing classes
+      # One day it'd be nice to parse existing classes
       @domain_classes = []
+      @frozen_domain_classes = []
       
       @output = ""
     end
-    
+
     def running?
       !!(@pipe)
     end
@@ -34,11 +35,42 @@ module Grails
       return @output.scan(/^Server running\. Browse to (.+)/).first[0]
     end
     
+    def add_domain_class(klass)
+      @domain_classes << klass unless @domain_classes.include?(klass)
+    end
+    
+    def needs_restart?
+      !running? || domain_changes?
+    end
+    
+    def domain_changes?
+      domain_changes.size > 0
+    end
+    
+    def domain_changes
+      (domain_classes - @frozen_domain_classes)
+    end
+    
+    def restart
+      log "Performing a restart."
+      
+      stop! if running?
+      run!
+    end
+    
     def run!
-      # We need to dump all our domain classes to actual files
-      domain_classes.each do |domain|
+      log "Currently loaded domain classes: #{@frozen_domain_classes}"
+      log "Need to load: #{domain_changes}"
+      
+      return unless domain_changes? || !running?
+      
+      # We need to dump all our _new or changed_ domain classes to actual files
+      domain_changes.each do |domain|
+        log "Writing #{domain.name} to disk..."
         domain.write_to @root
       end
+      
+      @frozen_domain_classes = domain_classes.clone
       
       install_inspector_controller_in(@root)
       
@@ -47,6 +79,7 @@ module Grails
       # Fancier:
       # @pipe, @pipe_out, @pipe_error = Open3.popen3("#{grails_executable} run-app", "r")
       
+      log "Starting Grails."
       @pipe = IO.popen("#{executable} run-app", "r")
       
       sleep(1)
@@ -61,17 +94,25 @@ module Grails
     end
     
     def stop!
-      Process.kill("TERM", @pipe.pid)
-      @pipe.close
-      @output = ""
+      if running?
+        log "Terminating Grails..."
       
-      @pipe = nil
+        Process.kill("TERM", @pipe.pid)
+        @pipe.close
+        @output = ""
+      
+        @pipe = nil
+      end
     end
     
     protected
     
     def update_output!
       @output << (@pipe.gets || "")
+    end
+    
+    def log(message)
+      puts "==> #{message}"
     end
   end
 end
